requirements.txt
fastapi[all]
uvicorn
pytest
requests
openpyxl
python-docx